ok actually exporting coverage data for source file trials

+++ src/org/adligo/tests4j/models/shared/metadata/I_TrialRunMetadata.java
@@ -39,7 +39,7 @@ public interface I_TrialRunMetadata {
 	 * which ran
 	 * @return
 	 */
-	public int getAllTestsCount();
+	public long getAllTestsCount();
 	
 	/**
 	 * returns the source file for .java files name which
