+++ src/org/adligo/tests4j/models/shared/metadata/TrialRunMetadataMutant.java
@@ -47,8 +47,8 @@ public class TrialRunMetadataMutant implements I_TrialRunMetadata {
 	}
 
 	@Override
-	public int getAllTestsCount() {
-		int toRet = 0;
+	public long getAllTestsCount() {
+		long toRet = 0;
 		for (I_TrialMetadata trial: trials) {
 			toRet = toRet + trial.getTestCount();
 		}
